Add in cron schedule for build, check for latest duckdb version and push new image if it doesn't exist

+++ docker_image.sh
@@ -4,15 +4,23 @@ platforms="linux/amd64,linux/arm64"
 
 echo "Getting latest tag version from duckdb"
 duckdb_version=$(git -C build/duckdb describe --tags --abbrev=0)
-echo "Building for duckdb version: ${duckdb_version}"
+image_name=datacatering/duckdb:"${duckdb_version}"
 
-docker run --privileged --rm tonistiigi/binfmt --install all
-docker buildx create --use --name builder
-docker buildx inspect --bootstrap builder
+if [ -z "$(docker images -q "$image_name" 2> /dev/null)" ]; then
+  echo "Building for duckdb version: ${duckdb_version}"
 
-docker buildx build \
-  --platform "$platforms" \
-  --build-arg "DUCKDB_VERSION=${duckdb_version}" \
-  -t datacatering/duckdb:"${duckdb_version}" --push .
+  docker run --privileged --rm tonistiigi/binfmt --install all
+  docker buildx create --use --name builder
+  docker buildx inspect --bootstrap builder
 
-echo "Done!"
+  docker buildx build \
+    --platform "$platforms" \
+    --build-arg "DUCKDB_VERSION=${duckdb_version}" \
+    -t "$image_name" --push .
+
+  echo "Done!"
+else
+  echo "Latest duckdb image version already exists, version: ${duckdb_version}"
+fi
+
+
